fix quiz card
